+++ src/AtomUI.Desktop.Controls/Splitter/Splitter.SemanticParts.cs
@@ -1,0 +1,25 @@
+using AtomUI.Controls.Primitives;
+using AtomUI.Theme;
+using Avalonia.Controls;
+
+namespace AtomUI.Desktop.Controls;
+
+[SemanticPart(
+    "dragger",
+    SelectorClass = "semantic-dragger",
+    SelectorRoute = "/template/ .semantic-scope-panel > .semantic-scope-handle /template/ .semantic-dragger",
+    ContractType = typeof(Thumb),
+    Cardinality = SemanticPartCardinality.Multiple,
+    Since = "6.0",
+    RuntimeCreated = true)]
+[SemanticPart(
+    "panel",
+    SelectorClass = "semantic-panel",
+    SelectorRoute = "/template/ .semantic-scope-panel > .semantic-panel",
+    ContractType = typeof(Control),
+    Cardinality = SemanticPartCardinality.Multiple,
+    Since = "6.0",
+    RuntimeCreated = true)]
+public partial class Splitter
+{
+}